Refactoring to use azure storage client adapter
