+++ Gemfile
@@ -1,4 +1,5 @@
 source 'https://rubygems.org'
 gemspec
-gem "azure-storage", "= 0.11.4.preview"
+gem 'azure-storage', '= 0.11.4.preview'
 gem 'rspec'
+gem 'coveralls', require: false
